+++ example/lib/services.dart
@@ -36,11 +36,10 @@ class MaxService extends BLoCService<TestBLoC> {
 	@override
 	void dispose() {
 		_counterSub.cancel();
-		print("DISPOSE");
 	}
 }
 
-class TriggeredService extends TriggerService<TestBLoC> {
+class TutorialService extends TriggerService<TestBLoC> {
 	@override
 	Future<void> trigger(TestBLoC bloc) async {
 		print(
@@ -54,32 +53,26 @@ class TriggeredService extends TriggerService<TestBLoC> {
 			"A BLoC consists of inputs, outputs, values, mappers and paramaters.    \n"
 			"Inputs and outputs must be controllers that provide Streams and Sinks  \n"
 			"e.g. StreamController                                                  \n"
-			"Optionally a provider and disposer can be created that provide helper  \n"
-			"classes. A provider is an Inherited widget that will make the BloC     \n"
-			"available by calling 'Provider.of(context)'.                           \n"
-			"A disposer is a Stateful widget that will automatically call dispose   \n"
-			"when the disposer is disposed by the flutter framework.                \n"
 			"                                                                       \n"
 		);
 		print(
 			"Mappers provide the bridge between inputs and outputs.                 \n"
 			"They are async functions that take in the input from the input         \n"
-			"stream and their return is added to the output stream.                 \n"
-			"If a mapper returns null it won't be added to the output stream.       \n"
+			"stream and any yeilded values are added to the output stream.          \n"
 			"                                                                       \n"
 		);
 		print(
-			"Values store the latest value on the BLoC for easy refrence,           \n"
-			"especially in mappers that want the last value from another output.    \n"
+			"Values store the latest value of an output on the BLoC for easy        \n"
+			"refrence, especially useful for mappers that want the last value from  \n"
+			"another output.                                                        \n"
 			"                                                                       \n"
-			"Paramaters are values that the BLoC receives on initialization and     \n"
-			"must be passed in the constructor to the BLoC or disposer.             \n"
+			"Paramaters are values that the BLoC receives on initialization.        \n"
 			"                                                                       \n"
 			"This print is in a service, it can do anything you want.               \n"
 			"                                                                       \n"
 		);
 		print(
-			"There are 4 types of services; input, output, BLoC and trigger.        \n"
+			"There are 5 types of services; input, output, BLoC, trigger and mapper.\n"
 			"Input and output take in a Stream or Sink, respective of their         \n"
 			"type, from the BLoC. They can then use it to modify the BLoC           \n"
 			"or receive updates from the BLoC.                                      \n"
@@ -88,12 +81,36 @@ class TriggeredService extends TriggerService<TestBLoC> {
 			"everything on the BLoC.                                                \n"
 			"                                                                       \n"
 			"Trigger services have access to the entire BLoC but must be triggerd by\n"
-			"other code that has access to the BLoC.                                \n"
+			"other code that has access to the BLoC e.g. through a provider.        \n"
+			"                                                                       \n"
+			"Mapper services are the same as a normal mapper but they are stored in \n"
+			"a seperate class so they can be reused and tested seperately.          \n"
+			"                                                                       \n"
+		);
+		print(
+			"A provider will make the BLoC available to all widgets below itself in \n"
+			"the widget tree. It can also retrieve the BLoC for you through the     \n"
+			"\"of\" method.                                                         \n"
+			"                                                                       \n"
+			"A disposer wraps a provider but will automatically call the dispose    \n"
+			"method for you when it goes out of context.                            \n"
+			"                                                                       \n"
+		);
+		print(
+			"Before running your app you must generate the BLoC. Information on this\n"
+			"can be found in the \"flutter_bloc_generator\" README.                 \n"
+			"                                                                       \n"
+			"The annotations are available in the \"flutter_bloc_annotations\"      \n"
+			"package, the generator commands are available in the                   \n"
+			"\"flutter_bloc_generator\" package and providers and disposers are     \n"
+			"available in the \"flutter_bloc_provider\" package.                    \n"
 			"                                                                       \n"
 		);
 	}
 }
 
 class StringifyMapper extends MapperService<int, String> {
-	String map(int inputData) => inputData.toString();
+	Stream<String> map(int inputData) async* {
+		yield inputData.toString();
+	}
 }